requirements.txt
Django
djangorestframework
youtube-search